Fix static issues for com.webos.service.location
:Release Notes:
Fix static issues for com.webos.service.location

:Detailed Notes:
Fix static issues for com.webos.service.location

:Testing Performed:
Build successfully

:QA Notes:
NA

:Issues Addressed:
[WRR-5299]- To fix static issues

Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.location/+/435600
Tested-by: muralidhar bn <muralidhar.n@lge.com>
Static-Analysis: <sanjay.b@lge.com>

+++ src/handler/position/gps/GpsWanInterface.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 LG Electronics, Inc.
+// Copyright (c) 2024 LG Electronics, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -239,8 +239,8 @@ bool GpsWanInterface::getContextCb(LSHandle *sh, LSMessage *message, void * cont
 
     jvalue_ref return_value_obj = {0};
     jvalue_ref getContext_obj = {0};
-    int context_count = 0;
-    int counter = 0;
+    long int context_count = 0;
+    long int counter = 0;
     char *context_string = NULL;
     bool return_value = false;
